code/requirements.txt
Flask==3.1.0
protobuf==6.30.2
python-dotenv==1.1.0
qdrant_client==1.14.2
sentence_transformers==4.1.0
google-genai==1.10.0